The queue_name should be left empty in most cases
If xivo-dird are listening to the same message bus for config changed events
and a queue_name is specified, the message will be received in round robin order.

Since we want to receive ALL messages that we are interested in, setting the queue_name
to the empty string will generate a new unique queue.